Starting integration of SDE code
This will introduce the L96s model as a unique model in the colllection
so that we can integrate some special features here first as a test
case.  However, integration.py can be minimally adjusted to rigorously
include the Euler-Maruyama and four-stage Runge-Kutta schemes for SDEs
with additive noise with minimal changes.