Second order Taylor scheme in rough draft form
Initial simplifications have been introduced to homogenize the code with
the rest of the repo.  The simple case of a Fourier truncation with p=1
is emphasized, as this is all that matters for order 2.0 convergence.
The general scheme discussed in the manuscript is not of particular
interest for DA so we neglect this implementation, included in that
manucript's repo.