Homogenizing with the Lorenz96 Equations
Using the stock Jacobian implementation in the extras file, changing
normal sampler to the standard univariate normal of appropriate
dimension for broadcasting.  Yet to be run on test cases, tomorrow will
most likely put additional time developing the model and testing it for
the overal integration with DAPPER for merge / push tutorial worksheets.