Loading initial example for SDE / perfect-random models
The initial results reproduce the basic conclusions of
Grudzien2020numerical as in the bib entry.  In short, we acheive
adequate performance with Ne=100, pert obs EnKF, no inflation or
localization when using a statistically robust SDE solver like
Runge-Kutta with dt=0.1, but we experience filter divergence
with the Euler-Maruyama scheme (also order 1.0) in the same setting.
Full analysis and a wider simulation range for parameters are in
the reference.