Initial example semi-complete
The interactive plot is set for a toy setting and parameter range is now
set in order to reproduce the reference.  Plotting is incomplete but it
seems like the results for the Euler-Maruyama are much more pessimistic
than in the reference when using a low-precision setting in a noise
driven dynamical system.  This may need additional debugging in
addition to finishing the plotting scripts.