Debugged and reproduced data for perfect random, pending plots
Need to work on plotting scripts still, but the data generation based on
combinator is working correctly now and has fully reproduced the results
of the reference.   This has now been added in the README along with a
few additional references that I think are worth mentioning.  Will work
on plotting scripts next week more than likely.