Updating TODO lists, changing some experiment outputs
Looks like the bug-fixes are all good to go, this does indeed closely
reproduce the manuscript results demonstrating the relationship between
model uncertainty and low-precision numerics, further validating the
twin experiment configuration that is recommended in the work.  I
changed some of the outputs to highlight what I think are the important
table entries, though the plotting is still unclear to me for this
script. I left the optimize field empty because of the lack of
inflation, etc. Otherwise, TODO fields seem to be taken care of,
but check if I understand the global seeding for the experiments
correctly.