Updating QG F90 typing
f2py was giving warnings / errors over the size of arrays exceeding stack sizes
on my cluser environment, yet I couldn't seem to find a way to change the
stack size with ulimit commands or to pass a recursive build option to
gcc through f2py as per the warning message. Changing the typing to
allocatable and allocating the arrays in their instance as per
https://www.scivision.dev/gfortran-stack-to-static-warning/
fixed these compile issues.